+++ src/extension.ts
@@ -36,6 +36,8 @@ class SidebarProvider implements vscode.WebviewViewProvider {
 		const scriptUri = webview.asWebviewUri(vscode.Uri.joinPath(this._extensionUri, 'media', 'main.js'));
 		const contentSecurityPolicy = [
 			"default-src 'none'",
+			"base-uri 'none'",
+			"form-action 'none'",
 			`img-src ${webview.cspSource}`,
 			`style-src ${webview.cspSource}`,
 			`script-src 'nonce-${nonce}'`